src/requirements.txt
bandit==1.7.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
filelock==3.7.1
Flask==2.1.3
Flask-Cors==3.0.10
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.23.1
pbr==5.9.0
py3-validate-email==1.0.7
pymongo==4.2.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
smmap==5.0.0
stevedore==4.0.0
threadpoolctl==3.1.0
urllib3==1.26.11
Werkzeug==2.2.0
xgboost==1.6.1
zipp==3.8.1
